@@ -1,14 +1,15 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import './index.css';
 import * as serviceWorker from './serviceWorkerRegistration';
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root')!);
+
+root.render(
     <React.StrictMode>
         <App />
     </React.StrictMode>,
-    document.getElementById('root')
 );
 
 // If you want your app to work offline and load faster, you can change
